Finalizando o estudo de estrutura de decisão!

+++ src/br/com/algoritimos/aula2/ProgramaPrincipal.java
@@ -6,7 +6,7 @@ public class ProgramaPrincipal {
 		
 		Calculadora calc1 = new Calculadora();
 		
-		int valorSomaRetornado = calc1.calcularSoma(10, 1);
+		Double valorSomaRetornado = calc1.calcularSoma(10.0, 1.0);
 		System.out.println(valorSomaRetornado);
 		
 		Double valorSubtracaoRetornado = calc1.calcularSubtracao(10.0, 1.0);
